Change layout & design for realm dialog
